backend/requirements.txt
flask
pandas
psycopg2-binary
flask_cors
cohere
openai
